refactor: CameraFragment, CameraUIView
In order to satisfy the SOLID principle, the inner class and interface were separated to the outside.

1. Detach the CameraUIEventsListener interface from within the CameraUIView interface to the outside(ISP)
2. Extract the CameraUIController from the CameraFragment class to the outside(SRP)

Now, When modifying features, it became clear which classes should be accessed.
